Quantum Computing Simulator/requirements.txt
numpy
pytest
matplotlib

PyQt5-Qt5-5.15.2
PyQt5-sip-12.13.0
pyqt5-5.15.10 
pyqtgraph-0.13.7